Implement actual web server
Well, with all this preparation the server is actually easy: read the
request, check if we can open the file, and send the file to the client.

There is one caveat here: we hardcode MIME type to be "text/plain". This
should be changed as the MIME type is important for browsers to display
web pages correctly.